+++ R/style.R
@@ -10,10 +10,12 @@
 #' * `call`: function calls
 #' * `string`: character literals
 #' * `comment`: comments
+#' * `bracket`: brackets: \code{(){}[]} 
 #'
 #' Each entry in a list must be a function that takes a character
-#' scalar, and returns a character scalar. The default style adds
-#' ANSI formatting to the code.
+#' scalar, and returns a character scalar with the exception of `bracket`
+#' which should be a list of these type of functions defining a color sequence. 
+#' The default style adds ANSI formatting to the code.
 #'
 #' Note that you can also change the code if you like, e.g. to include
 #' a unicode arrow character instead of the two-character assignment
@@ -26,6 +28,33 @@
 #' highlight(deparse(get), style = default_style())
 
 default_style <- function() {
+  rs <- rstudio$detect()
+  if (rs$type %in% c("rstudio_console", "rstudio_console_starting")) {
+    default_style_rstudio()
+  } else {
+    default_style_term()
+  }
+}
+
+default_style_rstudio <- function() {
+  theme <- rstudioapi::getThemeInfo()$editor
+  do.call(rstudio_themes[[theme]] %||% no_style, list())
+}
+
+no_style <- function() {
+  list(
+    reserved = identity,
+    number   = identity,
+    null     = identity,
+    operator = identity,
+    call     = identity,
+    string   = identity,
+    comment  = identity,
+    bracket  = list(identity)
+  )
+}
+
+default_style_term <- function() {
   list(
     reserved = red,
     number   = blue,
@@ -33,6 +62,7 @@ default_style <- function() {
     operator = green,
     call     = cyan,
     string   = yellow,
-    comment  = combine_styles(make_style("darkgrey"), italic)
+    comment  = combine_styles(make_style("darkgrey"), italic),
+    bracket  = c(yellow, blue, cyan)
   )
 }