+++ NEWS.md
@@ -1,5 +1,15 @@
 
-# 1.0.2
+# prettycode 1.1.0
+
+* Use the colors of the current theme in RStudio.
+
+* Color matching brackets to help distinguish between scopes
+  (#8, @mdequeljoe).
+
+* The withr package is now a development dependency, it is not needed
+  for regular use.
+
+# prettycode 1.0.2
 
 * Add `prettycode::prettycode()`. Call this function to turn on
   pretty-printing of function objects. This is needed to work around
@@ -8,11 +18,11 @@
 * If prettycode fails to highlight a function, it falls back to
   `base::print.function()` now (#3).
 
-# 1.0.1
+# prettycode 1.0.1
 
 * Avoid registering the `print.function` S3 method. This is needed to
   avoid a new `R CMD check` check
 
-# 1.0.0
+# prettycode 1.0.0
 
 First public release.